+++ src/libs/Detach/CHANGELOG.md
@@ -8,6 +8,7 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Added new sphere intersection method:
   - `Geometry3D.SphereTriangle`
+- Added `out float t` parameter to `Geometry3D.SphereCastSphere`.
 
 ## 0.29.0